+++ Paltarumi.Acopio.Balanza.Entity/Humedad.cs
@@ -5,10 +5,10 @@
         public int IdHumedad { get; set; }
         public int IdLote { get; set; }
         public DateTimeOffset Fecha { get; set; }
-        public float Tmh { get; set; }
-        public float Tms { get; set; }
-        public float Humedad100 { get; set; }
-        public float Humedad1 { get; set; }
+        public decimal Tmh { get; set; }
+        public decimal Tms { get; set; }
+        public decimal Humedad100 { get; set; }
+        public decimal Humedad1 { get; set; }
         public int CodigoMuestreoTipoMineral { get; set; }
         public int CodigoSupervisor { get; set; }
         public bool ReportadoProveedor { get; set; }
